requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
httpie==3.2.2
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
sqlparse==0.5.0
stripe==9.12.0
swagger-spec-validator==3.0.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
